fix: dispatch files in the correct folder
