+++ libs/coin-modules/coin-filecoin/jest.config.js
@@ -1,0 +1,9 @@
+/** @type {import('ts-jest/dist/types').JestConfigWithTsJest} */
+module.exports = {
+  collectCoverageFrom: ["src/**/*.ts"],
+  coverageDirectory: "coverage",
+  preset: "ts-jest",
+  testEnvironment: "node",
+  testPathIgnorePatterns: ["lib/", "lib-es/", ".integration.test.ts"],
+  modulePathIgnorePatterns: ["__tests__/fixtures"],
+};
