--- libs/coin-modules/coin-filecoin/src/currency.ts
@@ -1,11 +1,0 @@
-// import { CurrencyBridge } from "@ledgerhq/types-live";
-// import { getAccountShape } from "./utils/utils";
-// import { makeScanAccounts } from "@ledgerhq/coin-framework/bridge/jsHelpers";
-
-// const scanAccounts = makeScanAccounts({ getAccountShape });
-
-// export const currencyBridge: CurrencyBridge = {
-//   preload: () => Promise.resolve({}),
-//   hydrate: () => {},
-//   scanAccounts,
-// };
