+++ libs/coin-modules/coin-filecoin/src/index.ts
@@ -1,72 +1,1 @@
-import {
-  makeAccountBridgeReceive,
-  makeScanAccounts,
-  makeSync,
-} from "@ledgerhq/coin-framework/bridge/jsHelpers";
-import resolver from "./hw-getAddress";
-import getAddressWrapper from "@ledgerhq/coin-framework/bridge/getAddressWrapper";
-import { SignerContext } from "@ledgerhq/coin-framework/signer";
-import type { Account, AccountBridge, CurrencyBridge } from "@ledgerhq/types-live";
-import { defaultUpdateTransaction } from "@ledgerhq/coin-framework/bridge/jsHelpers";
-import type { Transaction, TransactionStatus } from "./types";
-import { getTransactionStatus } from "./getTransactionStatus";
-import { estimateMaxSpendable } from "./estimateMaxSpendable";
-import { prepareTransaction } from "./prepareTransaction";
-import { createTransaction } from "./createTransaction";
-import { getAccountShape } from "./utils/utils";
-import { buildSignOperation } from "./signOperation";
-import { broadcast } from "./broadcast";
-import { CoinConfig } from "@ledgerhq/coin-framework/config";
-import { FileCoinSigner } from "./types/signer";
-import fileCoinCurrencyConfig, { type FileCoinCurrencyConfig } from "./config";
-
-function buildCurrencyBridge(signerContext: SignerContext<FileCoinSigner>): CurrencyBridge {
-  const getAddress = resolver(signerContext);
-
-  const scanAccounts = makeScanAccounts({
-    getAccountShape,
-    getAddressFn: getAddressWrapper(getAddress),
-  });
-
-  return {
-    preload: () => Promise.resolve({}),
-    hydrate: () => {},
-    scanAccounts,
-  };
-}
-
-const sync = makeSync({ getAccountShape });
-
-function buildAccountBridge(
-  signerContext: SignerContext<FileCoinSigner>,
-): AccountBridge<Transaction, Account, TransactionStatus> {
-  const getAddress = resolver(signerContext);
-
-  const receive = makeAccountBridgeReceive(getAddressWrapper(getAddress));
-  const signOperation = buildSignOperation(signerContext);
-
-  return {
-    estimateMaxSpendable,
-    createTransaction,
-    updateTransaction: defaultUpdateTransaction,
-    getTransactionStatus,
-    prepareTransaction,
-    sync,
-    receive,
-    signOperation,
-    broadcast,
-  };
-}
-
-export function createBridges(
-  signerContext: SignerContext<FileCoinSigner>,
-  coinConfig: CoinConfig<FileCoinCurrencyConfig>,
-) {
-  fileCoinCurrencyConfig.setCoinConfig(coinConfig);
-
-  return {
-    currencyBridge: buildCurrencyBridge(signerContext),
-    accountBridge: buildAccountBridge(signerContext),
-  };
-}
-export { prepareTransaction, estimateMaxSpendable };
+export { createBridges } from "./bridge/index";
