+++ libs/coin-modules/coin-filecoin/src/network/index.ts
@@ -1,0 +1,1 @@
+export * from "./addresses";
