+++ libs/coin-modules/coin-filecoin/src/signer/getAddress.ts
@@ -3,10 +3,10 @@ import { log } from "@ledgerhq/logs";
 import { SignerContext } from "@ledgerhq/coin-framework/signer";
 import { GetAddressFn } from "@ledgerhq/coin-framework/bridge/getAddressWrapper";
 import { GetAddressOptions } from "@ledgerhq/coin-framework/derivation";
-import { getPath, isError } from "./utils";
-import { FileCoinSigner } from "./types/signer";
+import { getPath, isError } from "../bridge/utils";
+import { FilecoinSigner } from "../types";
 
-function resolver(signerContext: SignerContext<FileCoinSigner>): GetAddressFn {
+function resolver(signerContext: SignerContext<FilecoinSigner>): GetAddressFn {
   return async (deviceId: string, { path, verify }: GetAddressOptions) => {
     log("debug", "start getAddress process");
