+++ libs/coin-modules/coin-filecoin/src/signer/index.ts
@@ -1,0 +1,3 @@
+import resolver from "./getAddress";
+
+export default resolver;
