+++ libs/coin-modules/coin-filecoin/src/test/index.ts
@@ -1,0 +1,6 @@
+import makeCliTools from "./cli";
+
+export * from "./bridgeDatasetTest";
+export { makeCliTools };
+export * from "./bot-specs";
+export * from "./speculos-deviceActions";
