+++ libs/coin-modules/coin-filecoin/src/types/index.ts
@@ -1,0 +1,4 @@
+export * from "./bridge";
+export * from "./signer";
+export * from "./common";
+export * from "../errors";
