+++ libs/coin-modules/coin-filecoin/src/types/signer.ts
@@ -1,19 +1,19 @@
-export type FileCoinSignature = {
+export type FilecoinSignature = {
   return_code: number;
   error_message: string;
   signature_der: Uint8Array;
   signature_compact: Uint8Array;
 };
 
-export type FileCoinGetAddrResponse = {
+export type FilecoinGetAddrResponse = {
   addrByte: Uint8Array;
   addrString: string;
   compressed_pk: Uint8Array;
   return_code: number;
   error_message: string;
 };
-export interface FileCoinSigner {
-  showAddressAndPubKey(path: string): Promise<FileCoinGetAddrResponse>;
-  getAddressAndPubKey(path: string): Promise<FileCoinGetAddrResponse>;
-  sign(path: string, message: Uint8Array): Promise<FileCoinSignature>;
+export interface FilecoinSigner {
+  showAddressAndPubKey(path: string): Promise<FilecoinGetAddrResponse>;
+  getAddressAndPubKey(path: string): Promise<FilecoinGetAddrResponse>;
+  sign(path: string, message: Uint8Array): Promise<FilecoinSignature>;
 }
