+++ libs/ledger-live-common/src/families/filecoin/config.ts
@@ -1,2 +1,12 @@
-// Encapsulate for LLD & LLM
-export * from "@ledgerhq/coin-filecoin/config";
+import { ConfigInfo } from "@ledgerhq/live-config/LiveConfig";
+
+export const fileCoinConfig: Record<string, ConfigInfo> = {
+  config_currency_hedera: {
+    type: "object",
+    default: {
+      status: {
+        type: "active",
+      },
+    },
+  },
+};
